Refactor SHRINK-VECTOR
SHRINK-VECTOR was declared to be a macro that modified the place passed to it
as an argument. We modify it to be a function and inline %SHINK-VECTOR.

We modify one call side in lib/format.lisp in order to return the string
returned from SHRINK-VECTOR to adjust for the change.