Simplify IN-PACKAGE
The original IN-PACKAGE macro permitted quoted arguments and unquoted them
itself. Since this behaviour may allow for the user to overlook that they
quoted their argument to IN-PACKAGE, which in turn causes their code
portability with other Common Lisp implementations, we decide to make a
breaking change and cause IN-PACKAGE to signal an error instead. One
instance of a quoted package name in Cocoa IDE is corrected as a result of
this change.

Additionally, the original comment for IN-PACKAGE stated:

    This can only be partially backward-compatible: even if only
    the "name" arg is supplied, the old function would create the
    package if it didn't exist.

This is not consistent with behaviour of ANSI CL which states that IN-PACKAGE
"causes the the package named by name to become the current package - that is,
the value of *PACKAGE*. If no such package already exists, an error of type
PACKAGE-ERROR is signaled" and no package creation may happen. The function
SET-PACKAGE that is called by IN-PACKAGE also does not create packages when
invoked with names of non-existing packages. Therefore, we remove that comment
as it seems to no longer reflect reality.