Add CCL:STATIC-VARIABLE-P
The function returns true if the symbol is bound as a global static variable.
This complements BOUNDP since previously there was no exported way to check
if a defined special variable was dynamic or static, other than by attempting
to rebind it and checking if an error was signaled.

It is still required to make use of this information in DESCRIBE, INSPECT,
and APROPOS, and to modify CCL-TESTS to test this functionality.