Reimplement APROPOS
We remove the old APROPOS implementation and replace it with a simpler and
extensible one.

The new implementation is capable of recognizing when a symbol has a value
in the function, compiler macro function, macro function, special form,
keyword, constant variable, static variable, dynamic variable, condition,
or class namespaces. The status of these namespaces can be queried by querying
the *APROPOS-NAMESPACES* variable and new namespaces can be added via the
DEFINE-APROPOS-NAMESPACE macro. These two symbols are now exported from the
CCL package, documented, and overall available for public consumption.

We additionally clean up a single global variable from l1-init.lisp that
somehow was a part of the old definition of APROPOS.