Refactor VALID-CHAR-CODE{-P}
The function VALID-CHAR-CODE-P checked if its argument was a fixnum. We move
this check into the type VALID-CHAR-CODE and adjust the only other call site
of VALID-CHAR-CODE-P to use TYPEP instead of calling the function directly.
Furthermore, we rename VALID-CHAR-CODE-P to %VALID-CHAR-CODE-P to note that
it is a strictly internal function, and we move it from l1-aprims.lisp to
l1-typesys.lisp, next to its only current call site at DEFTYPE
VALID-CHAR-CODE.