Refactor AREF and ASET
The functions AREF and ASET exhibited code duplication. We factor out the
common part into macro %WITH-AREF-BODY and change AREF and ASET to call it.

A macro is used instead of a function for performance reasons - the resulting
code should be equivalent to the old one in order to avoid performance hits.