Clean up deprecated vstring and fixnum FASL encodings
This commit removes the deprecated vstring encoding from the whole FASL
machinery, including the FASL compiler, FASL loader, and cross-compiler FASL
loader. All strings are henceforth encoded using the nvstring encoding.

The following FASL opcodes and all support for them are removed as a result of
this change:
* $FASL-FIXNUM
* $FASL-VSTR
* $FASL-VMKSYM
* $FASL-VINTERN
* $FASL-VPKG-INTERN
* $FASL-VPKG
* $FASL-VINTERN-SPECIAL
* $FASL-VPKG-INTERN-SPECIAL
* $FASL-VMKSYM-SPECIAL

No attempt is made to clean up the resulting holes in FASL opcode numeration.
This exercise is left for the future commiters.

Additionally, this commit removes the old fixnum FASL encoding, since it is
no longer used in any code.

Additionally, the macro FASL-EPUSH-OP is moved from faslenv.lisp to
nfcomp.lisp since it is only utilized in the latter file.