Remove $FASL-GLOBALS from the FASL reader/writer
This commit removes parameters GNAMES and GOFFSETS from functions
FASL-{SCAN,DUMP-FILE,DUMP-BLOCK,DUMP} and the FASL operation $FASL-GLOBALS.

The parameters GNAMES and GOFFSETS accepted by and returned by the
aforementioned functions were always null. We therefore remove it as dead code
and remove the global variable *FASDUMP-GLOBAL-OFFSETS* that GOFFSETS was
bound to. Because the FASL operation $FASL-GLOBALS is now unused, we remove
it as well.