Remove FASL NVSTRING encoding
This commit removes the deprecated nvstring encoding from the whole FASL
machinery, including the FASL compiler, FASL loader, and cross-compiler FASL
loader. All strings are henceforth encoded using the vstring encoding.

The following FASL opcodes and all support for them are removed as a result of
this change:
* $FASL-NVSTR
* $FASL-NVMKSYM
* $FASL-NVINTERN
* $FASL-NVPKG-INTERN
* $FASL-NVPKG
* $FASL-NVINTERN-SPECIAL
* $FASL-NVPKG-INTERN-SPECIAL
* $FASL-NVMKSYM-SPECIAL

No attempt is made to clean up the resulting holes in FASL opcode numeration.
This exercise is left for the future commiters.